Switched the Quantity field from INT to DECIMAL(8, 2): - Updated the CREATE TABLE statement in CreateDatabase.sql. - Updated parameter conversion and result handling in Database.cs. - Updated InvoiceItem.cs field, property and Rehydrate parameters. - Updated InvoiceRenderer.cs to format Quantity using a format string that allows up to two decimal places but uses as few as possible. - Updated InvoiceRenderer.cs to measure the length of all the quantities and set the Qty column width correspondingly. Updated InvoiceRenderer.cs to only render the time component of Payment.ReceivedDateTime when it is non-zero. Added handling for payment types PaymentType.WireTransfer and PaymentType.PayPal to Payment.GetShortTypeDescription. Updated InvoiceRenderer.cs to not render DateTime.MinValue in the DueDate field. Updated dtpInvoiceDate and dtpDueDate to insert today's date when Home is pressed. Made DependencyProperty fields in InvoiceList.xaml.cs and InvoiceTabHeader.xaml.cs readonly.

+++ Invoices.Core/Database.cs
@@ -281,7 +281,7 @@ UPDATE Invoices
 
 				var sequenceParam = cmd.Parameters.Add("@Sequence", SqlDbType.Int);
 				var descriptionParam = cmd.Parameters.Add("@Description", SqlDbType.NVarChar);
-				var quantityParam = cmd.Parameters.Add("@Quantity", SqlDbType.Int);
+				var quantityParam = cmd.Parameters.Add("@Quantity", SqlDbType.Decimal);
 				var unitPriceParam = cmd.Parameters.Add("@UnitPrice", SqlDbType.Decimal);
 
 				for (int i=0; i < invoice.Items.Count; i++)
@@ -639,7 +639,7 @@ SELECT * FROM Invoices WHERE InvoiceID = @InvoiceID";
 		}
 	}
 
-	IEnumerable<(int InvoiceID, int Sequence, string Description, int Quantity, decimal UnitPrice)> ReadInvoiceItems(SqlDataReader reader)
+	IEnumerable<(int InvoiceID, int Sequence, string Description, decimal Quantity, decimal UnitPrice)> ReadInvoiceItems(SqlDataReader reader)
 	{
 		int invoiceID_ordinal = reader.GetOrdinal("InvoiceID");
 		int sequence_ordinal = reader.GetOrdinal("Sequence");
@@ -654,7 +654,7 @@ SELECT * FROM Invoices WHERE InvoiceID = @InvoiceID";
 			int invoiceID = reader.GetInt32(invoiceID_ordinal);
 			int sequence = reader.GetInt32(sequence_ordinal);
 			string description = reader.GetString(description_ordinal);
-			int quantity = reader.GetInt32(quantity_ordinal);
+			decimal quantity = reader.GetDecimal(quantity_ordinal);
 			decimal unitPrice = reader.GetDecimal(unitPrice_ordinal);
 
 			yield return (invoiceID, sequence, description, quantity, unitPrice);
